companies.import_siae: Transform methods into functions
