+++ itou/companies/management/commands/_import_siae/financial_annex.py
@@ -7,6 +7,7 @@ SiaeFinancialAnnex object logic used by the import_siae.py script is gathered he
 from django.utils import timezone
 
 from itou.companies.models import SiaeConvention, SiaeFinancialAnnex
+from itou.utils.python import timeit
 
 
 def get_creatable_and_deletable_afs(af_number_to_row):
@@ -81,3 +82,16 @@ def build_financial_annex_from_number(row):
             end_at=timezone.make_aware(row.end_at),
             convention=convention,
         )
+
+
+@timeit
+def manage_financial_annexes(af_number_to_row):
+    creatable_afs, deletable_afs = get_creatable_and_deletable_afs(af_number_to_row)
+
+    print(f"will create {len(creatable_afs)} financial annexes")
+    for af in creatable_afs:
+        af.save()
+
+    print(f"will delete {len(deletable_afs)} financial annexes")
+    for af in deletable_afs:
+        af.delete()
